@@ -19,46 +19,51 @@
   },
   "dependencies": {
     "@astrojs/mdx": "^2.1.1",
-    "@astrojs/react": "^3.0.9",
+    "@astrojs/react": "^3.0.10",
     "@astrojs/solid-js": "^4.0.1",
     "@astrojs/tailwind": "^5.1.0",
-    "@astrojs/vercel": "^7.1.1",
-    "@hookform/resolvers": "^3.1.0",
-    "@kobalte/core": "^0.9.6",
+    "@astrojs/vercel": "^7.3.0",
+    "@hookform/resolvers": "^3.3.4",
+    "@kobalte/core": "^0.12.1",
     "@motionone/solid": "^10.16.2",
-    "@nanostores/react": "^0.7.0",
-    "@nanostores/solid": "^0.3.2",
-    "@types/react": "^18.2.7",
-    "@types/react-dom": "^18.2.4",
-    "astro": "^4.3.2",
-    "astro-compress": "^2.2.9",
-    "axios": "^1.4.0",
-    "clsx": "^1.2.1",
-    "framer-motion": "^10.12.16",
-    "nanostores": "^0.9.3",
-    "node-html-parser": "^6.1.5",
-    "radash": "^10.8.1",
+    "@nanostores/react": "^0.7.1",
+    "@nanostores/solid": "^0.4.2",
+    "@radix-ui/react-slot": "^1.0.2",
+    "@types/react": "^18.2.55",
+    "@types/react-dom": "^18.2.19",
+    "astro": "^4.3.5",
+    "astro-compress": "^2.2.10",
+    "axios": "^1.6.7",
+    "clsx": "^2.1.0",
+    "dayjs": "^1.11.10",
+    "framer-motion": "^11.0.3",
+    "nanostores": "^0.9.5",
+    "node-html-parser": "^6.1.12",
+    "radash": "^11.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.43.9",
-    "react-use": "^17.4.0",
+    "react-hook-form": "^7.50.1",
+    "react-use": "^17.5.0",
     "react-use-measure": "^2.1.1",
     "remove-markdown": "^0.5.0",
-    "solid-js": "^1.7.5",
-    "tailwind-merge": "^1.12.0",
-    "zod": "^3.21.4",
-    "zustand": "^4.3.8"
+    "solid-js": "^1.8.14",
+    "tailwind-merge": "^2.2.1",
+    "tailwind-variants": "^0.2.0",
+    "vaul": "^0.9.0",
+    "zod": "^3.22.4",
+    "zustand": "^4.5.0"
   },
   "devDependencies": {
     "@hookform/devtools": "^4.3.1",
     "@ianvs/prettier-plugin-sort-imports": "^4.1.1",
-    "@tailwindcss/typography": "^0.5.9",
+    "@tailwindcss/container-queries": "^0.1.1",
+    "@tailwindcss/typography": "^0.5.10",
     "@total-typescript/ts-reset": "^0.5.1",
     "@types/eslint": "^8.56.2",
-    "@types/remove-markdown": "^0.3.1",
+    "@types/remove-markdown": "^0.3.4",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
-    "daisyui": "^2.51.6",
+    "csstype": "^3.1.3",
     "eslint": "^8.56.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^9.1.0",
@@ -72,30 +77,31 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-solid": "^0.13.1",
     "eslint-plugin-tailwindcss": "^3.14.2",
-    "eslint-plugin-vitest": "^0.3.21",
-    "husky": "^8.0.3",
-    "lint-staged": "^13.2.2",
+    "eslint-plugin-vitest": "^0.3.22",
+    "husky": "^9.0.10",
+    "lint-staged": "^15.2.2",
     "npm-run-all": "^4.1.5",
-    "postcss-html": "^1.5.0",
+    "postcss-html": "^1.6.0",
     "prettier": "^3.2.5",
     "prettier-plugin-astro": "^0.13.0",
     "prettier-plugin-tailwindcss": "^0.5.11",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-preset-minify": "^7.0.0",
-    "rehype-pretty-code": "^0.12.6",
+    "rehype-pretty-code": "^0.13.0",
     "rehype-slug": "^6.0.0",
-    "shiki": "^0.14.7",
+    "shiki": "^1.1.1",
     "shikiji": "^0.10.2",
     "stylelint": "^16.2.1",
     "stylelint-config-html": "^1.1.0",
-    "stylelint-config-recess-order": "^4.4.0",
+    "stylelint-config-recess-order": "^4.6.0",
     "stylelint-config-standard": "^36.0.0",
     "stylelint-declaration-block-no-ignored-properties": "^2.8.0",
     "stylelint-high-performance-animation": "^1.10.0",
     "stylelint-no-unsupported-browser-features": "^8.0.1",
     "stylelint-prettier": "^5.0.0",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.0.4"
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "^5.3.3"
   },
   "lint-staged": {
     "src/**/*.{css,js,ts,jsx,tsx,mdx,astro}": [
