@@ -1,58 +1,19 @@
 /** @jsxImportSource react */
+
 import { motion } from "framer-motion"
 import type { ComponentProps } from "react"
 
 import { cn } from "@/utils/helpers"
 
-export function AboutIcon({ className, ...rest }: ComponentProps<"svg">) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth={1.5}
-      stroke="currentColor"
-      className={cn("size-4", className)}
-      {...rest}
-    >
-      <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M17.982 18.725A7.488 7.488 0 0012 15.75a7.488 7.488 0 00-5.982 2.975m11.963 0a9 9 0 10-11.963 0m11.963 0A8.966 8.966 0 0112 21a8.966 8.966 0 01-5.982-2.275M15 9.75a3 3 0 11-6 0 3 3 0 016 0z"
-      />
-    </svg>
-  )
-}
-
-export function BlogIcon({ className, ...rest }: ComponentProps<"svg">) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth={1.5}
-      stroke="currentColor"
-      className={cn("size-4", className)}
-      {...rest}
-    >
-      <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M12 7.5h1.5m-1.5 3h1.5m-7.5 3h7.5m-7.5 3h7.5m3-9h3.375c.621 0 1.125.504 1.125 1.125V18a2.25 2.25 0 01-2.25 2.25M16.5 7.5V18a2.25 2.25 0 002.25 2.25M16.5 7.5V4.875c0-.621-.504-1.125-1.125-1.125H4.125C3.504 3.75 3 4.254 3 4.875V18a2.25 2.25 0 002.25 2.25h13.5M6 7.5h3v3H6v-3z"
-      />
-    </svg>
-  )
-}
-
-export function CheckIcon({ className, ...rest }: ComponentProps<"svg">) {
+export function CheckMotion({ className, ...rest }: ComponentProps<"svg">) {
   return (
     <svg
       fill="none"
       viewBox="0 0 24 24"
       stroke="currentColor"
       strokeWidth={3}
-      className={cn("size-4", className)}
       {...rest}
+      className={cn("size-5", className)}
     >
       <motion.path
         initial={{ pathLength: 0 }}
@@ -70,103 +31,53 @@ export function CheckIcon({ className, ...rest }: ComponentProps<"svg">) {
   )
 }
 
-export function MailIcon({ className, ...rest }: ComponentProps<"svg">) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={cn("size-4", className)}
-      {...rest}
-    >
-      <rect width="20" height="16" x="2" y="4" rx="2" />
-      <path d="m22 7-8.97 5.7a1.94 1.94 0 0 1-2.06 0L2 7" />
-    </svg>
-  )
-}
-
-export function RightChevronIcon({ className, ...rest }: ComponentProps<"svg">) {
+export function ChevronRight({ className, ...rest }: ComponentProps<"svg">) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 24 24"
-      strokeWidth={2}
+      strokeWidth={3}
       stroke="currentColor"
-      className={cn("size-4", className)}
       {...rest}
+      className={cn("size-5", className)}
     >
       <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
     </svg>
   )
 }
 
-export function PaperAirplaneIcon({ className, ...rest }: ComponentProps<"svg">) {
+export function Loader({ className, ...rest }: ComponentProps<"svg">) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      fill="none"
+      width="24"
+      height="24"
       viewBox="0 0 24 24"
-      strokeWidth={2}
+      fill="none"
       stroke="currentColor"
-      className={cn("size-4", className)}
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
       {...rest}
+      className={cn("size-5 animate-rotate", className)}
     >
-      <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M6 12L3.269 3.126A59.768 59.768 0 0121.485 12 59.77 59.77 0 013.27 20.876L5.999 12zm0 0h7.5"
-      />
+      <line x1="12" x2="12" y1="2" y2="6" className="opacity-[1]" />
+      <line x1="16.24" x2="19.07" y1="7.76" y2="4.93" className="opacity-[0.92]" />
+      <line x1="18" x2="22" y1="12" y2="12" className="opacity-[0.84]" />
+      <line x1="16.24" x2="19.07" y1="16.24" y2="19.07" className="opacity-[0.76]" />
+      <line x1="12" x2="12" y1="18" y2="22" className="opacity-[0.68]" />
+      <line x1="4.93" x2="7.76" y1="19.07" y2="16.24" className="opacity-[0.6]" />
+      <line x1="2" x2="6" y1="12" y2="12" className="opacity-[0.52]" />
+      <line x1="4.93" x2="7.76" y1="4.93" y2="7.76" className="opacity-[0.44]" />
     </svg>
   )
 }
 
-export function SnippetsIcon({ className, ...rest }: ComponentProps<"svg">) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth={1.5}
-      stroke="currentColor"
-      className={cn("size-4", className)}
-      {...rest}
-    >
-      <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M14.25 9.75L16.5 12l-2.25 2.25m-4.5 0L7.5 12l2.25-2.25M6 20.25h12A2.25 2.25 0 0020.25 18V6A2.25 2.25 0 0018 3.75H6A2.25 2.25 0 003.75 6v12A2.25 2.25 0 006 20.25z"
-      />
-    </svg>
-  )
-}
-
-export function SpinnerIcon({ className, ...rest }: ComponentProps<"svg">) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      className={cn("size-5 animate-spin text-slate-300", className)}
-      {...rest}
-    >
-      <circle
-        className="opacity-25"
-        cx="12"
-        cy="12"
-        r="10"
-        stroke="currentColor"
-        strokeWidth="4"
-      />
-      <path
-        className="opacity-75"
-        fill="currentColor"
-        d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-      />
-    </svg>
-  )
+export const Icons = {
+  chevronRight: ChevronRight,
+  loader: Loader,
+  animate: {
+    check: CheckMotion,
+  },
 }
