@@ -13,7 +13,7 @@ export default function Playground(props: {
   return (
     <div
       class={cn(
-        "relative rounded-xl border border-slate-400/20",
+        "relative rounded-md border",
         "bg-grid",
         "overflow-hidden",
         "aspect-[1/1] sm:aspect-[4/3] md:aspect-[10/6]",
