@@ -7,36 +7,41 @@ import { Icons } from "./icons"
 export default function Profile() {
   return (
     <div class="mx-auto max-w-3xl">
-      <div class="mb-12 flex shrink-0 items-baseline gap-x-6">
-        <div class="ml-1 flex size-[72px] items-center justify-center rounded-full bg-slate-900 text-slate-200 ring-2 ring-cyan-500">
-          <MotionLogo width={18} height={40.5} />
+      <div class="flex shrink-0 items-baseline gap-x-6">
+        <div class="flex size-[72px] items-center justify-center rounded-full bg-gradient text-heading">
+          <div class="flex size-[69px] items-center justify-center rounded-full bg-background">
+            <MotionLogo width={18} height={40.5} />
+          </div>
         </div>
         <div class="flex items-center gap-x-3">
           <h1 class="text-xl leading-5">kos3nz</h1>
           <a href="https://github.com/kos3nz" rel="noopener noreferrer" target="_blank">
-            <Icons.github class="size-5 text-slate-500 transition-colors duration-300 hover:text-cyan-500" />
+            <Icons.github class="size-5 text-inactive transition-colors duration-300 hover:text-primary" />
           </a>
           <a href="/contact/">
-            <Icons.mail class="size-5 text-slate-500 transition-colors duration-300 hover:text-cyan-500" />
+            <Icons.mail class="size-5 text-inactive transition-colors duration-300 hover:text-primary" />
           </a>
         </div>
       </div>
 
+      <hr class="my-8 h-px border-none bg-gradient-to-l from-transparent via-muted to-transparent" />
+
       <div class="space-y-10">
-        <div class="space-y-2 text-sm leading-6 text-slate-300">
-          <span class="block text-slate-500">About me</span>
+        <div class="space-y-2 text-sm leading-6">
+          <span class="block text-foreground/50">About me</span>
           <p>
             元ビール醸造士。独学でプログラミングを学び、普段はフロントエンドを中心に開発をしています。
             React, Typescript, Next.js, Remix, Solid.js などモダンな技術が好きです。
             より良いプロダクトを作れるよう、より良い開発者になれるよう、常に新しいことを学ぶことを心掛けています。
           </p>
         </div>
-        <div class="space-y-2 text-sm leading-6 text-slate-300">
-          <span class="block text-slate-500">Experienced</span>
+        <div class="space-y-2 text-sm leading-6">
+          <span class="block text-foreground/50">Experienced</span>
           <ul class="grid grid-cols-2 gap-2 md:grid-cols-3">
             <li>HTML</li>
             <li>CSS</li>
             <li>SCSS</li>
+            <li>styled-components</li>
             <li>TailwindCSS</li>
             <li>Javascript/Typescript</li>
             <li>Node.js</li>
@@ -47,7 +52,6 @@ export default function Profile() {
             <li>Astro</li>
             <li>tRPC</li>
             <li>Express</li>
-            <li>styled-components</li>
             <li>Redux</li>
             <li>Jest</li>
             <li>Storybook</li>
@@ -67,8 +71,8 @@ export default function Profile() {
             <li>GitHub</li>
           </ul>
         </div>
-        <div class="space-y-2 text-sm leading-6 text-slate-300">
-          <span class="block text-slate-500">Intrested</span>
+        <div class="space-y-2 text-sm leading-6">
+          <span class="block text-foreground/50">Intrested</span>
           <ul class="grid grid-cols-2 gap-2 md:grid-cols-3">
             <li>Rust</li>
             <li>React Native</li>
@@ -85,8 +89,8 @@ export default function Profile() {
             <li>AssemblyScript</li>
           </ul>
         </div>
-        <div class="space-y-2 text-sm leading-6 text-slate-300">
-          <span class="block text-slate-500">Tools</span>
+        <div class="space-y-2 text-sm leading-6">
+          <span class="block text-foreground/50">Tools</span>
           <ul class="grid grid-cols-2 gap-2 md:grid-cols-3">
             <li>VSCode</li>
             <li>Neovim</li>
@@ -95,8 +99,8 @@ export default function Profile() {
             <li>lazygit</li>
           </ul>
         </div>
-        <div class="space-y-2 text-sm leading-6 text-slate-300">
-          <span class="block text-slate-500">Environment</span>
+        <div class="space-y-2 text-sm leading-6">
+          <span class="block text-foreground/50">Environment</span>
           <ul class="grid grid-cols-2 gap-2 md:grid-cols-3">
             <li>Mac mini M2 Pro</li>
             <li>MacBook Pro - 2019</li>
