@@ -3,15 +3,17 @@
 import { For } from "solid-js"
 
 import SnippetCard from "@/components/solid/snippet-card"
-import { filterValue } from "@/store/filter-store.ts"
+import { filterValue } from "@/stores/filter.store"
 import type { Snippet } from "@/types/global.types.ts"
 
 export default function SnippetsList(props: { snippets: Snippet[] }) {
   const filtered = () => filterValue(props.snippets)
 
   return (
-    <div class="mt-12 grid gap-6 sm:grid-cols-2 md:gap-8 lg:grid-cols-3">
-      <For each={filtered()}>{(snippet) => <SnippetCard snippet={snippet} />}</For>
+    <div class="@container">
+      <div class="grid gap-5 @lg:grid-cols-2">
+        <For each={filtered()}>{(snippet) => <SnippetCard snippet={snippet} />}</For>
+      </div>
     </div>
   )
 }
