@@ -8,29 +8,30 @@ import { cn } from "@/utils/helpers"
 export default function TableOfContents(props: {
   headings: Heading[]
   activeHeading: Heading | undefined
+  cb?: () => void
 }) {
   return (
-    <div class="max-w-xs rounded-xl border border-slate-400/20 p-6 md:border-0">
-      <span class="block text-xs font-bold uppercase text-slate-300">On this page</span>
-      <ol class="mt-3 border-l-2 border-slate-400/20 text-xs font-semibold">
+    <div class="rounded-xl p-5">
+      <ol class="border-l-2 border-muted-foreground/20 text-[13px] font-medium">
         <For each={props.headings}>
           {(heading) => (
             <li class="leading-2 relative mt-3 first-of-type:mt-0">
               <a
                 href={heading.hash}
                 class={cn(
-                  "block leading-5 transition-colors duration-300",
+                  "block py-px leading-5 transition-colors duration-300",
                   (heading.depth === 1 || heading.depth === 2) && "pl-4",
                   heading.depth >= 3 && "pl-8",
                   props.activeHeading?.hash === heading.hash
-                    ? "text-slate-200"
-                    : "text-slate-500 hover:text-slate-200",
+                    ? "text-active"
+                    : "text-inactive hover:text-active",
                 )}
+                onClick={() => props.cb?.()}
               >
                 {heading.text}
               </a>
               <span
-                class="absolute left-[-2px] top-0 h-full w-[2px] origin-center bg-cyan-500 transition-transform duration-300"
+                class="absolute left-[-2px] top-0 h-full w-[2px] origin-center bg-primary transition-transform duration-300"
                 style={{
                   transform:
                     props.activeHeading?.hash === heading.hash
