@@ -1,13 +1,13 @@
 import { atom } from "nanostores"
 
-export const isDrawerOpen = atom(false)
+export const $isDrawerOpen = atom(false)
 
 export const openDrawer = () => {
-  isDrawerOpen.set(true)
+  $isDrawerOpen.set(true)
   document.documentElement.style.setProperty("--touch-action", "none")
 }
 
 export const closeDrawer = () => {
-  isDrawerOpen.set(false)
+  $isDrawerOpen.set(false)
   document.documentElement.style.setProperty("--touch-action", "auto")
 }
